lessons/180/python-app/requirements.txt
backoff==2.2.1
blinker==1.6.3
boto3==1.28.72
botocore==1.31.72
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Deprecated==1.2.14
Flask==3.0.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
packaging==23.2
prometheus-client==0.17.1
protobuf==4.24.4
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
psycopg2-binary==2.9.9
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.15.0
zipp==3.17.0
